+++ tests/corpus/samples/module/consistent-module/ConsistentModule.psm1
@@ -1,0 +1,9 @@
+function Get-ConsistentFoo {
+    param()
+    'foo'
+}
+function Set-ConsistentBar {
+    param()
+    'bar'
+}
+Export-ModuleMember -Function Get-ConsistentFoo, Set-ConsistentBar