+++ tests/corpus/samples/module/orphan-export/OrphanModule.psm1
@@ -1,0 +1,5 @@
+function Get-OrphanFoo {
+    param()
+    'foo'
+}
+Export-ModuleMember -Function Get-OrphanFoo